Tweak module and docs.
The most significant change is to parse the import argument options in
POSIX mode, so that they must come first. The incentive is that I can
(if I want) use the same parser for the subroutine arguments, so that I
can (e.g.) say
    load_module_ok '-no-load-error', 'My::Module';
should I want to turn the error diagnostic off for just the one call, or
    load_module_ok '-load-error', 'My::Module';
to turn it on.