Go back to a regex to recognize version numbers.
The rationale is that the version of the version module (sic) needed to
get is_lax() is much newer than the minimum Perl supported by the
Test2-Suite modules (and this one, for that matter). So I have tried to
minimize technical debt by adding an author test to detect a change in
the version::regex version of this regex.